Bugfix for reinstating a deleted user
When a user is deleted, they are not actually deleted. Their user object is hidden by the acts_as_paranoid gem. When
the same user tries to re-create their account, there was a bug where we were making a new user object with the same
UID. This fix will reinstate the "deleted" user insted of creating a new user object.